Fix: LedgerGlobalStats.collectionRate is Double; convert to Float for bar

+++ app/src/main/java/com/notification/app/ui/screens/FinancialStatsScreens.kt
@@ -71,7 +71,7 @@ fun LedgerStatsScreen(
             }
             item {
                 StatCard(if (isArabic) "نسبة التحصيل" else "Collection rate") {
-                    ProgressBarWithLabel(stats.collectionRate)
+                    ProgressBarWithLabel(stats.collectionRate.toFloat())
                 }
             }
             item {
